Refactor pipeline logic to use Strategy Pattern
